Fix cephadm bootstrap host delegation issues
Sometimes we see a bootstrap failure, where cephadm cannot read the
previously generated SSH key. This seems to be because it generates the
key on one host, then tries to bootstrap on another.

This can happen when running under Kayobe, and the ansible_host variable
is not respected with delegate_to.

There is another issue, where the facts seen when using delegate_to are
of the original host, but the task is executed on the delegate host.
This results in cephadm bootstrap getting the wrong value for --mon-ip,
and failing a port binding check.

This change reworks the cephadm role to avoid delegate_to, and instead
uses include_tasks and a condition, to only include for a single
bootstrap host.

Fixes: #62