Fix deployment with FQDN hostnames
cephadm demands that the name of the host given via ceph orch host add
equals the output of hostname on remote hosts [1].

In Ansible, the output of hostname is obtained with the nodename fact.
The hostname variable only returns the name until the first dot. See a
similar change in kolla-ansible for more details [2].

[1] https://docs.ceph.com/en/quincy/cephadm/host-management/#cephadm-fqdn
[2] https://review.opendev.org/c/openstack/kolla-ansible/+/660463