Pull ceph images with tag after docker login
In some environments issues are observed with images having missing tags
and this missing tag causing cephadm shell to resort to using the
default image. By pulling the ceph image to a host with the correct tag
cephadm shell should be able be able to detect the image and use it.